test -> runner : success

+++ src/main/java/me/kalpha/book/repository/AuthorRepository.java
@@ -4,8 +4,7 @@ import me.kalpha.book.entity.Author;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
-import java.math.BigInteger;
 
 @RepositoryRestResource(collectionResourceRel = "authors", path = "authors")
-public interface AuthorRepository extends JpaRepository<Author, BigInteger>  {
+public interface AuthorRepository extends JpaRepository<Author, Long>  {
 }
