+++ src/main/java/me/kalpha/book/repository/BookAuthorRepository.java
@@ -4,8 +4,6 @@ import me.kalpha.book.entity.BookAuthor;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
-import java.math.BigInteger;
-
 @RepositoryRestResource(collectionResourceRel = "bookAuthors", path = "bookAuthors")
-public interface BookAuthorRepository extends JpaRepository<BookAuthor, BigInteger> {
+public interface BookAuthorRepository extends JpaRepository<BookAuthor, Long> {
 }
